+++ social_core/pipeline/partial.py
@@ -1,13 +1,6 @@
 from functools import wraps
 
 
-def save_status_to_session(strategy, pipeline_index, *args, **kwargs):
-    """Saves current social-auth status to session."""
-    strategy.session_set('partial_pipeline',
-                         strategy.partial_to_session(pipeline_index + 1,
-                                                     *args, **kwargs))
-
-
 def partial(func):
     @wraps(func)
     def wrapper(strategy, pipeline_index, *args, **kwargs):