+++ social_core/tests/actions/actions.py
@@ -145,7 +145,6 @@ class BaseActionTest(unittest.TestCase):
                 'social_core.pipeline.social_auth.social_details',
                 'social_core.pipeline.social_auth.social_uid',
                 'social_core.pipeline.social_auth.auth_allowed',
-                'social_core.pipeline.partial.save_status_to_session',
                 'social_core.tests.pipeline.ask_for_password',
                 'social_core.pipeline.social_auth.social_user',
                 'social_core.pipeline.user.get_username',